Create README.md
The program I have developed is based on a list that allows you to organize and plan your daily tasks. I have used Js, Html, and css technologies to create it. I hope this tool is very useful for you.

Improvement:

I have created a program that relies on a dynamic list to help you sort and plan out your daily tasks. This tool has been crafted using cutting-edge Js, Html, and css technologies. I sincerely hope that it proves to be an invaluable asset in your life